# Conflicts:
#	CHANGELOG.md
#	CMakeLists.txt
#	SofaKernel/framework/sofa/core/behavior/BaseConstraintCorrection.cpp
#	SofaKernel/framework/sofa/defaulttype/TopologyTypes.h
#	SofaKernel/framework/sofa/helper/ComponentChange.cpp
#	SofaKernel/modules/SofaRigid/JointSpring.inl
#	SofaKernel/modules/SofaSimulationGraph/SimpleApi.cpp
#	modules/SofaGeneralAnimationLoop/MechanicalMatrixMapper.cpp
#	modules/SofaGeneralAnimationLoop/MechanicalMatrixMapper.h
#	modules/SofaGeneralEngine/NearestPointROI.cpp
#	modules/SofaGraphComponent/SceneCheckerListener.cpp
#	modules/SofaGraphComponent/SceneCheckerListener.h
#	modules/SofaHaptics/src/SofaHaptics/LCPForceFeedback.cpp

+++ SofaKernel/modules/SofaSimpleFem/TetrahedronDiffusionFEMForceField.inl
@@ -1,6 +1,6 @@
 /******************************************************************************
 *       SOFA, Simulation Open-Framework Architecture, development version     *
-*                (c) 2006-2018 INRIA, USTL, UJF, CNRS, MGH                    *
+*                (c) 2006-2019 INRIA, USTL, UJF, CNRS, MGH                    *
 *                                                                             *
 * This program is free software; you can redistribute it and/or modify it     *
 * under the terms of the GNU Lesser General Public License as published by    *
@@ -37,7 +37,7 @@
 
 #include <iostream>
 #include <fstream>
-#include <string.h>
+#include <cstring>
 
 #include <SofaSimulationTree/GNode.h>
 #include <sofa/helper/AdvancedTimer.h>
@@ -192,14 +192,14 @@ void TetrahedronDiffusionFEMForceField<DataTypes>::init()
     if(tetraDiff.size()==0)
     {
         tetraDiff.resize(topology->getNbTetrahedra());
-        for(int i=0; i<topology->getNbTetrahedra(); i++)
+        for(size_t i=0; i<topology->getNbTetrahedra(); i++)
             tetraDiff[i]=this->d_constantDiffusionCoefficient.getValue();
     }
     /// possible input tetrahedral diffusion coefficient
     else
     {
         loadedDiffusivity = true;
-        if((int)(tetraDiff.size())!=topology->getNbTetrahedra())
+        if(tetraDiff.size() != topology->getNbTetrahedra())
         {
             msg_error() <<"Wrong size of potential input vector.";
             return;
@@ -267,7 +267,7 @@ template <class DataTypes>
 typename TetrahedronDiffusionFEMForceField<DataTypes>::Real TetrahedronDiffusionFEMForceField<DataTypes>::getTetraDiffusionCoefficient(unsigned int i)
 {
     sofa::helper::vector<Real> tetraDiff = this->d_tetraDiffusionCoefficient.getValue();
-    if((int) i <= topology->getNbTetrahedra())
+    if(i <= topology->getNbTetrahedra())
     {
         return tetraDiff[i];
     }
@@ -291,7 +291,7 @@ void TetrahedronDiffusionFEMForceField<DataTypes>::setDiffusionCoefficient(const
     sofa::helper::vector <Real>& tetraDiff = *(d_tetraDiffusionCoefficient.beginEdit());
     tetraDiff.clear();
     tetraDiff.resize(topology->getNbTetrahedra());
-    for(int i=0; i<topology->getNbTetrahedra(); i++)
+    for(size_t i=0; i<topology->getNbTetrahedra(); i++)
         tetraDiff[i] = d_constantDiffusionCoefficient.getValue();
     d_tetraDiffusionCoefficient.endEdit();
 
@@ -309,7 +309,7 @@ void TetrahedronDiffusionFEMForceField<DataTypes>::setDiffusionCoefficient(const
 
     // Save the new diffusion coefficient for each tetrahedron
     sofa::helper::vector <Real>& tetraDiff = *(d_tetraDiffusionCoefficient.beginEdit());
-    for(int i=0; i<topology->getNbTetrahedra(); i++)
+    for(size_t i=0; i<topology->getNbTetrahedra(); i++)
         tetraDiff[i] = val[i];
     d_tetraDiffusionCoefficient.endEdit();
